backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.30.0
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6
